perf: splited two func
